# Improve npm install and CI tests times