--- a/buildScripts/build.js
+++ b/buildScripts/build.js
@@ -1,28 +1,45 @@
 const exec = require('child_process').execSync;
 const rimraf = require('rimraf');
-const fs = require('fs');
+const fs = require('fs-extra');
 const path = require('path');
+const TASKS_UTILS_DIR = 'artifactory-tasks-utils';
+const TASKS_DIR = 'tasks';
+const TESTS_DIR = 'tests';
 
-cleanExecNpm('pack', 'artifactory-tasks-utils');
+installArtifactoryTaskUtils();
 installTasks();
-cleanExecNpm('i', 'tests');
+installTests();
+
+/**
+ *
+ * Install Artifactory task utils.
+ */
+function installArtifactoryTaskUtils() {
+    clean(TASKS_UTILS_DIR, true);
+    execNpm('i', TASKS_UTILS_DIR);
+    execNpm('pack', TASKS_UTILS_DIR);
+}
 
 /**
  * Install tasks.
  */
 function installTasks() {
-    fs.readdir('tasks', (err, files) => {
+    fs.readdir(TASKS_DIR, (err, files) => {
         files.forEach(taskName => {
-            let taskDir = path.join('tasks', taskName);
+            let taskDir = path.join(TASKS_DIR, taskName);
             // If a package.json is missing, npm will exec the install command on the parent folder. This will cause an endless install loop.
             if (fs.existsSync(path.join(taskDir, 'package.json'))) {
-                cleanExecNpm('i', taskDir);
+                clean(taskDir);
+                copyTaskUtilsModules(taskDir);
+                execNpm('i', taskDir);
             } else {
                 fs.readdir(taskDir, (err, taskVersionDirs) => {
                     taskVersionDirs.forEach(versToBuild => {
                         let taskVersionDir = path.join(taskDir, versToBuild);
                         if (fs.existsSync(path.join(taskVersionDir, 'package.json'))) {
-                            cleanExecNpm('i', taskVersionDir);
+                            clean(taskVersionDir);
+                            copyTaskUtilsModules(taskVersionDir);
+                            execNpm('i', taskVersionDir);
                         }
                     });
                 });
@@ -32,21 +49,40 @@ function installTasks() {
 }
 
 /**
- * Clean npm install/pack files.
- * @param cwd - (String) - Current working directory.
+ * Install tests.
  */
-function clean(cwd) {
-    rimraf.sync(path.join(cwd, 'node_modules'));
-    rimraf.sync(path.join(cwd, 'package-lock.json'));
-    rimraf.sync(path.join(cwd, '*.tgz'));
+function installTests() {
+    clean(TESTS_DIR);
+    copyTaskUtilsModules(TESTS_DIR);
+    execNpm('i', TESTS_DIR);
 }
 
 /**
- * Clean directory and execute npm command.
+ * Copy artifactory-tasks-utils/node_modules to dest/node_modules
+ * @param dest - The destination
+ */
+function copyTaskUtilsModules(dest) {
+    fs.copySync(path.join(TASKS_UTILS_DIR, 'node_modules'), path.join(dest, 'node_modules'));
+}
+
+/**
+ * Clean npm install/pack files.
+ * @param cwd - (String) - Current working directory.
+ * @param cleanPackage (Boolean) - True to clean the 'npm pack' results.
+ */
+function clean(cwd, cleanPackage) {
+    rimraf.sync(path.join(cwd, 'node_modules'));
+    rimraf.sync(path.join(cwd, 'package-lock.json'));
+    if (cleanPackage) {
+        rimraf.sync(path.join(cwd, '*.tgz'));
+    }
+}
+
+/**
+ *
  * @param command - (String) - The command to execute, i.e. install, pack, etc.
  * @param cwd - (String) - Current working directory.
  */
-function cleanExecNpm(command, cwd) {
-    clean(cwd);
-    exec('npm ' + command + ' -q', { cwd: cwd, stdio: [0, 1, 2] });
+function execNpm(command, cwd) {
+    exec('npm ' + command + ' -q --no-fund', { cwd: cwd, stdio: [0, 1, 2] });
 }