--- a/buildScripts/package.json
+++ b/buildScripts/package.json
@@ -4,12 +4,13 @@
     "author": "JFrog",
     "private": true,
     "license": "Apache-2.0",
-    "dependencies": {
+    "devDependencies": {
         "edit-json-file": "^1.1.0",
         "command-line-args": "^5.0.2",
         "command-line-usage": "^5.0.5",
         "compare-versions": "^3.4.0",
         "assert": "^1.4.1",
-        "rimraf": "^2.6.2"
+        "rimraf": "^2.6.2",
+        "fs-extra": "^7.0.0"
     }
 }