* Reduce `npm install` time by copying the shared dependencies of artifactory-task-utils to each one of the tasks. Installation of each task will avoid downloading this dependencies.
* Clean up build log.
* Currently the CI tests run twice - once before creating the VSIX package and once before running the tests. Now we will run `npm i` automatically only before tests.

**Benchmarks**

Benchmark on my machine, running `time npm i`
* Before: `270.79s user 35.66s system 116% cpu 4:22.39 total` 
* After: `62.13s user 16.68s system 74% cpu 1:45.50 total`

Benchmark on Appveyor:
* Before: https://ci.appveyor.com/project/yahavi/artifactory-azure-devops-extension/builds/32905150
  * Windows: `17 min 35 sec`
  * Ubuntu:  `9 min 37 sec`

* After: https://ci.appveyor.com/project/yahavi/artifactory-azure-devops-extension/builds/32917562
  * Windows: `12 min 30 sec`
  * Ubuntu: ` 8 min 48 sec`